refactor(translate): gather the translate services under one directory
PdfTranslationService sat at the services root while the text path lived in
services/translate/, so one feature carried two layouts. Move it in, and give
the class files the PascalCase names the naming conventions ask for.

Drop TranslateService.releaseWindow and TranslateTask.ownedBy: nothing calls
them, and window death is already covered by the task's own `destroyed` watch.
