+++ src/main/services/translate/PdfTranslationService.ts
