+++ src/main/services/translate/__tests__/TranslateService.test.ts
@@ -42,7 +42,7 @@ vi.mock('../../../ai/streamManager/listeners/WebContentsListener', () => ({
 }))
 
 const { makeModel } = await import('../../../ai/__tests__/fixtures')
-const { TranslateService } = await import('../translateService')
+const { TranslateService } = await import('../TranslateService')
 const translateService = new TranslateService()
 
 const TARGET: TranslateLanguage = {
